Add Apollo Engine caching
Apollo engine caching defaults to the lowest-defined caching
setting it finds (default is 0).
This means that we have to set an explicit caching setting on each
of our types if we want generic caching behaviour.
(ie, if we didn't set it on "walls", any query which included walls
would never be cached)

I've set the time to 90 seconds but that's pretty arbitrary.

Evidence of caching can be seen
- locally by removing the 'cacheControl' from
  `strip: ['cacheControl', ... ] in the Apollo Engine setup
- by logging into Apollo Engine itself, which gives a visual
  timeline of requests, and shows how many were returned from
  the cache

'stores' config was [pinched from the Request Caching docs](https://www.apollographql.com/docs/engine/caching.html#engine-cache-config)