Overriding definitions of GraphQL default scalar types
Same as with the schema, I'm passing the i18n object into the
resolver function and then translating the *copy_pasted*
definitions :sad: that I took from the default graphql scalar
types.
I was looking for a smarter way to do this, but I didn't think
of any.
If ever graphql changes its definitions, ours will be out of
sync.