Throw error if both 'next' and 'previous' are passed in
This doesn't technically fail without us throwing an error, but
it does return a surprising result (ie, it ignores one parameter).
It's weird behaviour, so let's get rid of it.

Another thing is that returning an Error from a GraphQL resolver
will still return a 200 status code (ie, usually means the request
is "okay"). This is a design choice by the GraphQL people so I've
put it in the test just to be explicit.

[Source](https://stackoverflow.com/a/42942539)