Add API field documentation
Since GraphQL can introspect itself (win!!), there are
a few libraries that can quickly generate docs for us.

graphql-markdown generates a markdown file listing the
various types that are returned by the API.

To generate the docs:
- the API must be running locally
- `yarn docs`

It's possible that the README stuff will get restructured
as right now it's very "get the app started" focused, but
for now this is a reasonable place for these docs to live.